requirements.txt
PyQt5>=5.15.7
python-docx>=0.8.11
openpyxl>=3.1.2
